Move builder related functions into a separate Builder module.
